Varying input sizes; randomness in input. When the distribution is high; the compression ratio is low. When there is more repetition; compression ratio is very high; example 178 mb to 1 mb file.

+++ compressionalgorithms/src/main/java/org/bbolla/algorithms/compression/lossless/DemoStringCompression.java
@@ -101,7 +101,7 @@ public class DemoStringCompression {
     private static void logCompressionStats(String input, String compressed) {
         Integer inputSize = input.getBytes().length;
         Integer compressedSize = compressed.getBytes().length;
-        double ratio = ((double) inputSize) / ((double) compressedSize);
+        double ratio = (double) inputSize / compressedSize;
         log.info("Compression ratio : {}", ratio);
     }
